chore(conversation): move ConversationsListVirtual component
