+++ src/components/LeftSidebar/ConversationsList/ConversationsListVirtual.vue
@@ -37,8 +37,8 @@
 <script>
 import { RecycleScroller } from 'vue-virtual-scroller'
 
-import Conversation from './ConversationsList/Conversation.vue'
-import LoadingPlaceholder from '../LoadingPlaceholder.vue'
+import Conversation from './Conversation.vue'
+import LoadingPlaceholder from '../../LoadingPlaceholder.vue'
 
 import 'vue-virtual-scroller/dist/vue-virtual-scroller.css'
